# docs/api/openapi.yml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Oulib API Documentation",
    "description": "API documentation for Oulib - the library management system",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/borrowing": {
      "post": {
        "tags": [
          "Borrowing"
        ],
        "summary": "Tạo phiếu mượn sách",
        "description": "Thủ thư tạo bản ghi mượn cho người dùng theo danh sách barcode và thời hạn mượn",
        "operationId": "borrowBook",
        "requestBody": {
          "description": "Thông tin mượn sách gồm người mượn, thời hạn và danh sách barcode",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BorrowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Tạo phiếu mượn thành công",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "400": {
            "description": "Yêu cầu mượn không hợp lệ",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "401": {
            "description": "Chưa xác thực",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "403": {
            "description": "Không có quyền hoặc vi phạm điều kiện mượn",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy người dùng hoặc sách",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "409": {
            "description": "Barcode bị trùng hoặc đã có bản ghi mượn",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi hệ thống",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/borrowing/return": {
      "post": {
        "tags": [
          "Borrowing"
        ],
        "summary": "Trả sách",
        "description": "Thủ thư thực hiện trả sách theo danh sách barcode của bản sao cần trả",
        "operationId": "returnBook",
        "requestBody": {
          "description": "Danh sách barcode bản sao sách cần trả",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReturnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Trả sách thành công",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "400": {
            "description": "Yêu cầu trả sách không hợp lệ",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "401": {
            "description": "Chưa xác thực",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "403": {
            "description": "Không có quyền thực hiện",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy sách hoặc bản ghi mượn",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi hệ thống",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/borrowing/{userId}/history": {
      "get": {
        "tags": [
          "Borrowing"
        ],
        "summary": "Lấy lịch sử mượn theo người dùng",
        "description": "Thủ thư hoặc quản trị viên xem lịch sử mượn/trả của một người dùng cụ thể",
        "operationId": "getUserBorrowingHistory",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID của người dùng cần xem lịch sử",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Trạng thái lọc: borrowing, return/returned, overdue",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy lịch sử mượn thành công",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "400": {
            "description": "Giá trị trạng thái lọc không hợp lệ",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "401": {
            "description": "Chưa xác thực",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "403": {
            "description": "Không có quyền thực hiện",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy người dùng",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi hệ thống",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/borrowing/records": {
      "get": {
        "tags": [
          "Borrowing"
        ],
        "summary": "Lấy danh sách bản ghi mượn",
        "description": "Trả về danh sách bản ghi mượn có phân trang và hỗ trợ lọc theo trạng thái, người mượn",
        "operationId": "getBorrowRecords",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Trạng thái lọc: borrowing, return/returned, overdue",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "borrowerId",
            "in": "query",
            "description": "ID người mượn cần lọc",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Số trang, bắt đầu từ 0",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "Số phần tử mỗi trang",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy danh sách bản ghi mượn thành công",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseBorrowRecordResponse"
                }
              }
            }
          },
          "400": {
            "description": "Tham số lọc hoặc phân trang không hợp lệ",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseBorrowRecordResponse"
                }
              }
            }
          },
          "401": {
            "description": "Chưa xác thực",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseBorrowRecordResponse"
                }
              }
            }
          },
          "403": {
            "description": "Không có quyền thực hiện",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseBorrowRecordResponse"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi hệ thống",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseBorrowRecordResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/borrowing/records/{recordId}": {
      "get": {
        "tags": [
          "Borrowing"
        ],
        "summary": "Lấy chi tiết bản ghi mượn",
        "description": "Trả về thông tin chi tiết của một bản ghi mượn theo recordId",
        "operationId": "getBorrowRecordDetail",
        "parameters": [
          {
            "name": "recordId",
            "in": "path",
            "description": "ID của bản ghi mượn cần xem",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy chi tiết bản ghi thành công",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBorrowRecordDetailResponse"
                }
              }
            }
          },
          "401": {
            "description": "Chưa xác thực",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBorrowRecordDetailResponse"
                }
              }
            }
          },
          "403": {
            "description": "Không có quyền thực hiện",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBorrowRecordDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy bản ghi mượn",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBorrowRecordDetailResponse"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi hệ thống",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBorrowRecordDetailResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/borrowing/history": {
      "get": {
        "tags": [
          "Borrowing"
        ],
        "summary": "Lấy lịch sử mượn của tôi",
        "description": "Người dùng hiện tại xem lịch sử mượn/trả của chính mình, có thể lọc theo trạng thái",
        "operationId": "getMyBorrowingHistory",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Trạng thái lọc: borrowing, return/returned, overdue",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy lịch sử mượn thành công",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "400": {
            "description": "Giá trị trạng thái lọc không hợp lệ",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "401": {
            "description": "Chưa xác thực",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy người dùng",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi hệ thống",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponseObject": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "type": "object"
          }
        }
      },
...
      "BorrowRequest": {
        "type": "object",
        "description": "DTO yêu cầu BorrowRequest",
        "properties": {
          "borrowerId": {
            "type": "string",
            "description": "Trường borrowerId",
            "example": "id-001"
          },
          "borrowDuration": {
            "type": "integer",
            "format": "int32",
            "description": "Trường borrowDuration",
            "example": 1,
            "minimum": 0
          },
          "barcodes": {
            "type": "array",
            "description": "Trường barcodes",
            "example": [],
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "borrowDuration"
        ]
      },
      "ReturnRequest": {
        "type": "object",
        "description": "DTO yêu cầu ReturnRequest",
        "properties": {
          "barcodes": {
            "type": "array",
            "description": "Trường barcodes",
            "example": [],
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "ApiResponsePageResponseBorrowRecordResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/PageResponseBorrowRecordResponse"
          }
        }
      },
      "BorrowRecordResponse": {
        "type": "object",
        "description": "DTO phản hồi BorrowRecordResponse",
        "properties": {
          "id": {
            "type": "string",
            "description": "Trường id",
            "example": "id-001"
          },
          "borrowerId": {
            "type": "string",
            "description": "Trường borrowerId",
            "example": "id-001"
          },
          "barcode": {
            "type": "string",
            "description": "Trường barcode",
            "example": "BC-0001"
          },
          "bookTitle": {
            "type": "string",
            "description": "Trường bookTitle",
            "example": "Clean Code"
          },
          "thumbnailUrl": {
            "type": "string",
            "description": "Trường thumbnailUrl",
            "example": "https://example.com/resource"
          },
          "borrowDate": {
            "type": "string",
            "format": "date",
            "description": "Trường borrowDate",
            "example": "2026-04-08"
          },
          "dueDate": {
            "type": "string",
            "format": "date-time",
            "description": "Trường dueDate",
            "example": "2026-04-08T10:00:00"
          },
          "returnDate": {
            "type": "string",
            "format": "date",
            "description": "Trường returnDate",
            "example": "2026-04-08"
          },
          "status": {
            "type": "string",
            "description": "Trường status",
            "enum": [
              "BORROWING",
              "RETURNED",
              "OVERDUE"
            ],
            "example": "BORROWING"
          }
        }
      },
      "PageResponseBorrowRecordResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BorrowRecordResponse"
            }
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "ApiResponseBorrowRecordDetailResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/BorrowRecordDetailResponse"
          }
        }
      },
      "BorrowRecordDetailResponse": {
        "type": "object",
        "description": "DTO phản hồi BorrowRecordDetailResponse",
        "properties": {
          "id": {
            "type": "string",
            "description": "Trường id",
            "example": "id-001"
          },
          "borrowerId": {
            "type": "string",
            "description": "Trường borrowerId",
            "example": "id-001"
          },
          "borrowerFullName": {
            "type": "string",
            "description": "Trường borrowerFullName",
            "example": "Nguyen Van A"
          },
          "borrowerEmail": {
            "type": "string",
            "description": "Trường borrowerEmail",
            "example": "user@example.com"
          },
          "barcode": {
            "type": "string",
            "description": "Trường barcode",
            "example": "BC-0001"
          },
          "borrowDate": {
            "type": "string",
            "format": "date-time",
            "description": "Trường borrowDate",
            "example": "2026-04-08T10:00:00Z"
          },
          "dueDate": {
            "type": "string",
            "format": "date",
            "description": "Trường dueDate",
            "example": "2026-04-08"
          },
          "status": {
            "type": "string",
            "description": "Trường status",
            "enum": [
              "BORROWING",
              "RETURNED",
              "OVERDUE"
            ],
            "example": "BORROWING"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}